re-add dependencies and comment for dependency clarity

+++ Engine/Source/Graphics/Pipelines/MeshPipeline.cpp
@@ -195,7 +195,7 @@ void MeshPipeline::UpdateProjectionViewModelUniform(LLGL::CommandBuffer& command
 	meshSettings.textureClip = glm::mat4(1.0f);
 	meshSettings.hasBones[0] = mesh.mHasBones;
 
-	meshSettings.hasBones[1] = mesh.mCurrentBoneIndex;
+	meshSettings.hasBones[1] = static_cast<float>(mesh.mCurrentBoneIndex);
 
 	std::uint64_t transformOffset = 0;
 	for (const auto& finalTransform : mesh.mFinalTransforms)
